+++ src/main/java/com/zincoid/nullbot/core/component/ai/chat/client/BaseAiClient.java
@@ -1,6 +1,6 @@
 package com.zincoid.nullbot.core.component.ai.chat.client;
 
-import com.zincoid.nullbot.core.component.ai.chat.memory.ChatMemory;
+import com.zincoid.nullbot.core.component.ai.chat.memory.Memory;
 import com.zincoid.nullbot.core.component.ai.chat.message.BaseMessage;
 import com.zincoid.nullbot.core.component.ai.chat.message.Message;
 import com.zincoid.nullbot.core.component.ai.chat.model.Model;
@@ -15,19 +15,19 @@ import java.util.List;
 @RequiredArgsConstructor
 public class BaseAiClient implements AiClient<BaseMessage> {
 
-    private final ChatMemory chatMemory;
+    private final Memory memory;
     private final Model model;
 
     @Override
     public BaseMessage call(String chatId, String prompt, BaseMessage message, boolean thinking, int maxTokens) {
-        chatMemory.add(chatId, message);
-        List<Message> messages = chatMemory.get(chatId);
+        memory.add(chatId, message);
+        List<Message> messages = memory.get(chatId);
         List<Message> _messages = new ArrayList<>();
         _messages.add(BaseMessage.system(prompt));
         _messages.addAll(messages);
         ModelResponse response = model.invoke(_messages, thinking, maxTokens);
         BaseMessage _message = BaseMessage.assistant(response.getContent());
-        chatMemory.add(chatId, _message);
+        memory.add(chatId, _message);
         return _message;
     }
 }